Updated Tags, Post, and User Table, Rep, Service and Controller.

+++ src/main/java/com/example/demo/Services/Concrete/UserService.java
@@ -26,7 +26,7 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public User GetUserByUserId(long userId) {
+    public User GetUserByUserId(int userId) {
         return _userRepository.findById(userId).orElse(null);
     }
